feat: correct project files
