+++ src/NetBlog.Aplication/NetBlog.Aplication.csproj
